More warning cleanup #605

+++ libraries/chain/database.cpp
@@ -1269,12 +1269,12 @@ void database::update_witness_schedule4()
    modify( wso, [&]( witness_schedule_object& _wso )
    {
       // active witnesses has exactly STEEMIT_MAX_WITNESSES elements, asserted above
-      for( int i = 0; i < active_witnesses.size(); i++ )
+      for( size_t i = 0; i < active_witnesses.size(); i++ )
       {
          _wso.current_shuffled_witnesses[i] = active_witnesses[i];
       }
 
-      for( int i = active_witnesses.size(); i < STEEMIT_MAX_WITNESSES; i++ )
+      for( size_t i = active_witnesses.size(); i < STEEMIT_MAX_WITNESSES; i++ )
       {
          _wso.current_shuffled_witnesses[i] = account_name_type();
       }
@@ -1425,12 +1425,12 @@ void database::update_witness_schedule()
             _wso.current_shuffled_witnesses.push_back( w );
             */
          // active witnesses has exactly STEEMIT_MAX_WITNESSES elements, asserted above
-         for( int i = 0; i < active_witnesses.size(); i++ )
+         for( size_t i = 0; i < active_witnesses.size(); i++ )
          {
             _wso.current_shuffled_witnesses[i] = active_witnesses[i];
          }
 
-         for( int i = active_witnesses.size(); i < STEEMIT_MAX_WITNESSES; i++ )
+         for( size_t i = active_witnesses.size(); i < STEEMIT_MAX_WITNESSES; i++ )
          {
             _wso.current_shuffled_witnesses[i] = account_name_type();
          }
@@ -3126,7 +3126,7 @@ try {
       modify( get_feed_history(), [&]( feed_history_object& fho )
       {
          fho.price_history.push_back( median_feed );
-         int steemit_feed_history_window = STEEMIT_FEED_HISTORY_WINDOW_PRE_HF_16;
+         size_t steemit_feed_history_window = STEEMIT_FEED_HISTORY_WINDOW_PRE_HF_16;
          if( has_hardfork( STEEMIT_HARDFORK_0_16__551) )
             steemit_feed_history_window = STEEMIT_FEED_HISTORY_WINDOW;
 
@@ -3973,7 +3973,7 @@ void database::set_hardfork( uint32_t hardfork, bool apply_now )
 {
    auto const& hardforks = get_hardfork_property_object();
 
-   for( int i = hardforks.last_hardfork + 1; i <= hardfork && i <= STEEMIT_NUM_HARDFORKS; i++ )
+   for( uint32_t i = hardforks.last_hardfork + 1; i <= hardfork && i <= STEEMIT_NUM_HARDFORKS; i++ )
    {
       if( i <= STEEMIT_HARDFORK_0_5__54 )
          _hardfork_times[i] = head_block_time();
